3.0 FEATURE SET #19: generalized beam
NEW SPANNER FUNCTIONALITY: generalized abjad.Beam (closes #319).

    NEW:

        * abjad.Beam.beam_lone_notes
        * abjad.Beam.beam_rests
        * abjad.Beam.durations
        * abjad.Beam.span_beam_count
        * abjad.Beam.stemlet_length

    REMOVED:

        * abjad.ComplexBeam
        * abjad.DuratedComplexBeam
        * abjad.MeasuredComplexBeam
        * abjad.MultipartBeam

SPANNER CHANGES:

    OLD: HorizontalBracketSpanner
    NEW: HorizontalBracket

REMOVED SPANNER: abjad.ComplexTrillSpanner; use abjad.TrillSpanner instead.

MARKUP CHANGE: removed markup courtesy autocolumn (closes #977).

    OLD:

        >>> staff = abjad.Staff("c'4 d' e' f'")
        >>> abjad.attach(abjad.Markup('Allegro'), staff[0])
        >>> abjad.attach(abjad.Markup('non troppo'), staff[0])

        >>> abjad.f(staff)
        \new Staff
        {
            c'4
            - \markup {
                \column
                    {
                        \line
                            {
                                Allegro
                            }
                        \line
                            {
                                "non troppo"
                            }
                    }
                }
            d'4
            e'4
            f'4
        }

    NEW:

        >>> staff = abjad.Staff("c'4 d' e' f'")
        >>> abjad.attach(abjad.Markup('Allegro'), staff[0])
        >>> abjad.attach(abjad.Markup('non troppo'), staff[0])

        >>> abjad.f(staff)
        \new Staff
        {
            c'4
            - \markup { Allegro }
            - \markup { "non troppo" }
            d'4
            e'4
            f'4
        }

    * Also removed abjad.Markup.stack_priority property
