+++ abjad/core/test/test_Leaf__split_by_durations.py
@@ -648,9 +648,9 @@ def test_Leaf__split_by_durations_20():
     staff.append(abjad.Measure((2, 8), "c'8 d'8"))
     staff.append(abjad.Measure((2, 8), "e'8 f'8"))
     leaves = abjad.select(staff).leaves()
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, leaves[:2])
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, leaves[-2:])
     slur = abjad.Slur()
     abjad.attach(slur, leaves)
@@ -724,9 +724,9 @@ def test_Leaf__split_by_durations_21():
     staff.append(abjad.Measure((2, 8), "c'8 d'8"))
     staff.append(abjad.Measure((2, 8), "e'8 f'8"))
     leaves = abjad.select(staff).leaves()
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, leaves[:2])
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, leaves[-2:])
     slur = abjad.Slur()
     abjad.attach(slur, leaves)
@@ -797,7 +797,7 @@ def test_Leaf__split_by_durations_22():
     """
 
     note = abjad.Note("c'8")
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, abjad.select(note))
 
     assert format(note) == "c'8\n[\n]"
@@ -824,9 +824,9 @@ def test_Leaf__split_by_durations_23():
     staff.append(abjad.Measure((2, 8), "c'8 d'8"))
     staff.append(abjad.Measure((2, 8), "e'8 f'8"))
     leaves = abjad.select(staff).leaves()
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, leaves[:2])
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, leaves[-2:])
     slur = abjad.Slur()
     abjad.attach(slur, leaves)
