+++ abjad/spanners/__init__.py
@@ -1,15 +1,10 @@
 from .Spanner import Spanner
 from .Beam import Beam
 from .BowContactSpanner import BowContactSpanner
-from .ComplexBeam import ComplexBeam
-from .ComplexTrillSpanner import ComplexTrillSpanner
-from .DuratedComplexBeam import DuratedComplexBeam
 from .Glissando import Glissando
 from .Hairpin import Hairpin
-from .HorizontalBracketSpanner import HorizontalBracketSpanner
-from .MeasuredComplexBeam import MeasuredComplexBeam
+from .HorizontalBracket import HorizontalBracket
 from .MetronomeMarkSpanner import MetronomeMarkSpanner
-from .MultipartBeam import MultipartBeam
 from .OctavationSpanner import OctavationSpanner
 from .PhrasingSlur import PhrasingSlur
 from .PianoPedalSpanner import PianoPedalSpanner
