+++ abjad/spanners/test/test_Beam__fracture.py
@@ -37,7 +37,7 @@ def test_Beam__fracture_01():
 def test_Beam__fracture_02():
 
     staff = abjad.Staff("c'8 d'8 e'8 f'8 g'8 a'8 b'8 c''8")
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, staff[:4])
     beam._fracture(1, direction=abjad.Left)
 
@@ -133,7 +133,7 @@ def test_Beam__fracture_05():
     """
 
     staff = abjad.Staff("c'8 cs'8 d'8 ef'8 e'8 f'8 fs'8 g'8")
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, staff[:5])
     beam._fracture(2, direction=None)
 
@@ -168,7 +168,7 @@ def test_Beam__fracture_06():
     """
 
     staff = abjad.Staff("c'8 cs'8 d'8 ef'8 e'8 f'8 fs'8 g'8")
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, staff[:5])
     beam._fracture(0, direction=None)
 
@@ -201,7 +201,7 @@ def test_Beam__fracture_07():
     """
 
     staff = abjad.Staff("c'8 cs'8 d'8 ef'8 e'8 f'8 fs'8 g'8")
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, staff[:5])
     beam._fracture(4, direction=None)
 
@@ -234,7 +234,7 @@ def test_Beam__fracture_08():
     """
 
     staff = abjad.Staff("c'8 cs'8 d'8 ef'8 e'8 f'8 fs'8 g'8")
-    beam = abjad.Beam()
+    beam = abjad.Beam(beam_lone_notes=True)
     abjad.attach(beam, staff[:5])
     beam._fracture(-1, direction=None)
