--- abjad/spanners/test/test_ComplexBeam___eq__.py
@@ -1,12 +1,0 @@
-import abjad
-
-
-def test_ComplexBeam___eq___01():
-    """
-    Spanner is strict comparator.
-    """
-
-    spanner_1 = abjad.ComplexBeam()
-    spanner_2 = abjad.ComplexBeam()
-
-    assert not spanner_1 == spanner_2
