--- abjad/spanners/test/test_MultipartBeam___init__.py
@@ -1,10 +1,0 @@
-import abjad
-
-
-def test_MultipartBeam___init___01():
-    """
-    Initialize empty multipart beam spanner.
-    """
-
-    beam = abjad.MultipartBeam()
-    assert isinstance(beam, abjad.MultipartBeam)
