+++ abjad/system/LilyPondFormatManager.py
@@ -200,29 +200,9 @@ class LilyPondFormatManager(AbjadObject):
             down_markup_wrappers,
             neutral_markup_wrappers,
             ):
-            if not wrappers:
-                continue
-            elif 1 < len(wrappers):
-                direction = wrappers[0].indicator.direction
-                if direction is None:
-                    direction = '-'
-                wrappers = wrappers[:]
-                wrappers.sort(key=lambda _: -_.indicator.stack_priority)
-                lines = [
-                    abjad.Markup.line(
-                        [_.indicator],
-                        deactivate=_.deactivate,
-                        tag=_.tag,
-                        )
-                    for _ in wrappers
-                    ]
-                markup = abjad.Markup.column(lines, direction=direction)
-                format_pieces = markup._get_format_pieces()
-                bundle.right.markup.extend(format_pieces)
-            else:
-                wrapper = wrappers[0]
+            for wrapper in wrappers:
                 if wrapper.indicator.direction is None:
-                    markup = abjad.Markup(wrappers[0].indicator, direction='-')
+                    markup = abjad.Markup(wrapper.indicator, direction='-')
                 else:
                     markup = wrapper.indicator
                 format_pieces = markup._get_format_pieces()
